add benchmark for multiple partition send and receive
