+++ namespace_test.go
@@ -3,18 +3,14 @@ package eventhub
 import (
 	"context"
 	"flag"
-	"fmt"
 	mgmt "github.com/Azure/azure-sdk-for-go/services/eventhub/mgmt/2017-04-01/eventhub"
 	rm "github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2017-05-10/resources"
 	"github.com/Azure/go-autorest/autorest/adal"
 	"github.com/Azure/go-autorest/autorest/azure"
 	log "github.com/sirupsen/logrus"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 	"math/rand"
 	"os"
-	"pack.ag/amqp"
-	"sync"
 	"testing"
 	"time"
 )
@@ -73,87 +69,6 @@ func (suite *eventHubSuite) SetupSuite() {
 
 func (suite *eventHubSuite) TearDownSuite() {
 	// tear down queues and subscriptions maybe??
-}
-
-func (suite *eventHubSuite) TestBasicOperations() {
-	tests := map[string]func(*testing.T, *Namespace, *mgmt.Model){
-		"TestSend":           testBasicSend,
-		"TestSendAndReceive": testBasicSendAndReceive,
-	}
-
-	ns := suite.getNamespace()
-
-	for name, testFunc := range tests {
-		setupTestTeardown := func(t *testing.T) {
-			hubName := randomName("goehtest", 10)
-			mgmtHub, err := ns.EnsureEventHub(context.Background(), hubName)
-			defer ns.DeleteEventHub(context.Background(), hubName)
-
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			testFunc(t, ns, mgmtHub)
-		}
-
-		suite.T().Run(name, setupTestTeardown)
-	}
-}
-
-func testBasicSend(t *testing.T, ns *Namespace, mgmtHub *mgmt.Model) {
-	hub, err := ns.NewEventHub(*mgmtHub.Name)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	err = hub.Send(context.Background(), &amqp.Message{
-		Data: []byte("Hello!"),
-	})
-	assert.Nil(t, err)
-}
-
-func testBasicSendAndReceive(t *testing.T, ns *Namespace, mgmtHub *mgmt.Model) {
-	partitionID := (*mgmtHub.PartitionIds)[0]
-	hub, err := ns.NewEventHub(*mgmtHub.Name, HubWithPartitionedSender(partitionID))
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer hub.Close()
-
-	numMessages := rand.Intn(100) + 20
-	var wg sync.WaitGroup
-	wg.Add(numMessages + 1)
-
-	messages := make([]string, numMessages)
-	for i := 0; i < numMessages; i++ {
-		messages[i] = randomName("hello", 10)
-	}
-
-	go func() {
-		for idx, message := range messages {
-			ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-			err := hub.Send(ctx, &amqp.Message{Data: []byte(message)}, SendWithMessageID(fmt.Sprintf("%d", idx)))
-			cancel()
-			if err != nil {
-				log.Println(idx)
-				log.Fatalln(err)
-			}
-		}
-		defer wg.Done()
-	}()
-
-	count := 0
-	err = hub.Receive(partitionID, func(ctx context.Context, msg *amqp.Message) error {
-		assert.Equal(t, messages[count], string(msg.Data))
-		count++
-		wg.Done()
-		return nil
-	}, ReceiveWithPrefetchCount(100))
-	if err != nil {
-		t.Fatal(err)
-	}
-	wg.Wait()
-
 }
 
 func (suite *eventHubSuite) ensureProvisioned(tier mgmt.SkuTier) error {
